requirements.txt
Flask==0.12.2
Keras==2.1.6
Pillow==5.2.0
qrcode==6.0
gunicorn==19.9.0
numpy==1.14.3
tensorflow==1.8.0
